requirements.txt
beautifulsoup4==4.6.3
lxml==4.4.1
numpy==1.17.2
pandas==0.25.1
python-edgar==2.6.1
urllib3==1.24.3